Finished issue #13
Created "Store" POJO - A simple POJO representing an Store.
Created "StoreAPI", method getStores - An API to retrieve JSON data and store it as a List of "Store".
Created "StoreAPI", method getStore - An API to retrieve JSON data and store it as a "Store".
Refactor rename: "OffersAPI" to "OfferAPI".
Created "OfferAPI's", method getOffer - An API to retrieve JSON data and store it as a "Offer".